moved inspect_string_custom_excluding_substring into separate file
